pyproject.toml
[tool.poetry]
name = "ferrea"
version = "2.1.0"
description = "Module for cross-cutting concerns designed for Ferrea application."
authors = ["Eugenio Grimoldi"]
license = "MIT"
readme = "README.md"
packages = [{include = "ferrea", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
neo4j = "^5.28.1"
geopy = "^2.3.0"
attrs = "^25.3.0"
fastapi = "^0.115.12"
pyyaml-include = "^2.2"
pydantic-core = "^2.33.2"
pydantic = "^2.11.4"
loguru = "^0.7.3"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.5"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "./src"
